Moved hd wallet functions from fsm
